ocr/requirements.txt
pillow>=10.0.0
pdfplumber>=0.11.0
pypdfium2>=4.30.0
pytesseract>=0.3.10
